refactor: writeMarkdownFile のファイルを独立させる
